Resolvido posicionamento dos componentes do PerfilHeader
